pyproject.toml
[project]
name = "openproject-timesheet-export"
version = "1.0.0"
description = "Export OpenProject time entries to formatted Excel timesheets"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["openproject", "timesheet", "export", "excel"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
    "openpyxl>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/jeremi/openproject-timesheet-export"
Repository = "https://github.com/jeremi/openproject-timesheet-export"
Issues = "https://github.com/jeremi/openproject-timesheet-export/issues"

[project.scripts]
openproject-export = "export_open_project:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["export_open_project.py"]

[tool.uv]
dev-dependencies = []
